+++ CHANGELOG.md
@@ -27,7 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New config param 'resourceNameOverride' to support OpenAPI specifications with no tags.
-
+### Changed
+- Config param 'generateResourceClasses' renamed to 'generateResourceInterfaces'.
+ 
 ## [1.1.0] - 2025-05-20
 
 ### Added